MaxwellPDEs

Documentation for MaxwellPDEs.

MaxwellPDEs.Absorber

MaxwellPDEs.AbsorberParameters

MaxwellPDEs.AbstractSusceptibility

MaxwellPDEs.PML

MaxwellPDEs.PMLParameters

MaxwellPDEs.Simulation

MaxwellPDEs.StablePML

MaxwellPDEs.Structure

MaxwellPDEs.Absorber — Type

Absorber

Specifies an adiabatic absorber along a single dimension. This absorber is well-suited for dispersive materials intersecting with absorbing edges of the simulation at the expense of more layers.

Examples

pml = Absorber(num_layers=40)

MaxwellPDEs.AbsorberParameters — Type

AbsorberParameters

This struct is used to store the sigma parameters of the PML.

Examples

absorber_params = AbsorberParameters(sigma_order=3, sigma_min=0.0, sigma_max=1.5)

Fields

sigma_order: The order of the sigma function.

sigma_min: The minimum value of the sigma function.

sigma_max: The maximum value of the sigma function.

MaxwellPDEs.AbstractSusceptibility — Type

AbstractSusceptibility

A susceptibility is used to build dispersion models.

A dispersion model is a collection of susceptibilities of the same type or different types.

The supported susceptibilities LorentzianSusceptibility DrudeSusceptibility PoleResidue

Each susceptibility returns a symmetric 3*3 matrix.

The sigma field is used when the susceptibility is isotropic

The sigmadiag and sigmaoffdiag vectors are used when the susceptibility is anisotropic and can

be used to turn on/off the anisotropic terms.

MaxwellPDEs.PML — Type

PML

Specifies a standard PML along a single dimension.

Fields

num_layers: The number of PML layers

pml_params: Parameters of the complex frequency-shifted absorption poles

absorber_params: Parameters to fine tune the absorber profile and properties

Examples

pml = PML(num_layers=10)

MaxwellPDEs.PMLParameters — Type

PMLParameters

This struct is used to store the kappa and alpha parameters of the PML.

Examples

pmal_params = PMLParameters(sigma_order=3, sigma_min=0.0, sigma_max=1.5, kappa_min=0.0)

Fields

kappa_order: The order of the kappa function.

kappa_min: The minimum value of the kappa function.

kappa_max: The maximum value of the kappa function.

alpha_order: The order of the alpha function.

alpha_min: The minimum value of the alpha function.

alpha_max: The maximum value of the alpha function.

absorber_params: The parameters of the PML absorber.

MaxwellPDEs.Simulation — Type

Simulation

Contains all data needed by mesher and em-solver to run a simulation.

The simulation object shouldn't include any information about space or time discretization or any information related to the solver or numerical algorithms used. Its only purpose is to hold all the data needed by the solver and mesher to run a simulation.

The workflow is as follows:

Construct all simulation componenets

Call the mesh function to get a mesh object

Discretize the simulation data

Initalize fields for storing the solution

Call the solve function to get the solution

MaxwellPDEs.StablePML — Type

StablePML

Specifies a 'stable' PML along a single dimension. This PML deals handles possbly divergent simulations better, but at the expense of more layers.

Example

 pml = StablePML(num_layers=40)

MaxwellPDEs.Structure — Type

Structure{GeometryType}

Defines a physical object with a shape and medium.

Fields

geometry: Defines the geometry of the structure.

medium: Defines the material of the structure.