MaxwellPDEs

Documentation for MaxwellPDEs.

MaxwellPDEs.AbstractSusceptibility

MaxwellPDEs.Simulation

MaxwellPDEs.Structure

MaxwellPDEs.AbstractSusceptibility — Type

AbstractSusceptibility

A susceptibility is used to build dispersion models.

A dispersion model is a collection of susceptibilities of the same type or different types.

The supported susceptibilities LorentzianSusceptibility DrudeSusceptibility PoleResidue

Each susceptibility returns a symmetric 3*3 matrix.

The sigma field is used when the susceptibility is isotropic

The sigmadiag and sigmaoffdiag vectors are used when the susceptibility is anisotropic and can

be used to turn on/off the anisotropic terms.

MaxwellPDEs.Simulation — Type

Simulation

Contains all data needed by mesher and em-solver to run a simulation.

The simulation object shouldn't include any information about space or time discretization or any information related to the solver or numerical algorithms used. Its only purpose is to hold all the data needed by the solver and mesher to run a simulation.

The workflow is as follows:

Construct all simulation componenets

Call the mesh function to get a mesh object

Discretize the simulation data

Initalize fields for storing the solution

Call the solve function to get the solution

MaxwellPDEs.Structure — Type

Structure{GeometryType}

Defines a physical object with a shape and medium.

Fields

geometry: Defines the geometry of the structure.

medium: Defines the material of the structure.